log scales for string charts with range > 2 orders of magnitude
@salspaugh this is a bit of a departure from the previous version so i
wanted to see what you thought of it before implementing it for other
datatypes.

had to use a dot plot because it’s not really kosher to use bars on a
log scale … which is good because it creates visual distinctiveness to
make it less likely they will be confused for linear scales, but also
introduces a new chart type that’s quite different from what was tested.